clipboard-apis/integration_with_execcommand_events_are_syncronous_copy_on_iframecontentwindow.html

<!DOCTYPE html>
<html><head>
    <meta charset="UTF-8">
    <title>Clipboard API - integration with execCommand, events are syncronous  - copy on 'iframe.contentWindow'</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/_lib.js"></script>
    <script>
/** integration with execCommand, events are syncronous  */
/* Events: copy cut */
/* Test HTML: <iframe id="iframe"></iframe>*/
/* Targets: 'iframe.contentWindow' */
/* This assertation is based on Gecko's implementation. */
var eventFired = false;
function clipboard_api_test(e){
    /* the event is syncronous, must not run after the execCommand call is finished */
    assert_false(window.afterExecCommand);
    eventFired = true;
}
function onTestSetupReady( eventType ){
  document.getElementsByTagName('p')[0].firstChild.data='FAILED'; // if event doesn't fire, this test failed..
    var ifr=document.getElementById('iframe'), doc;
    (doc=ifr.contentDocument).designMode='on';
    doc.body.innerHTML='<p>Frame contents</p>';
    ifr.contentWindow.focus();
    var rng=doc.createRange();
    rng.selectNodeContents(doc.getElementsByTagName('p')[0]);
    var sel=ifr.contentWindow.getSelection();
    sel.removeAllRanges();
    sel.addRange(rng);
}
function triggerTestManually(eventType){
    var ifr=document.getElementById('iframe');
    window.afterExecCommand=false;
    /* This triggers the event listener, which asserts false */
    var execSucceeded = ifr.contentDocument.execCommand(eventType, false, false);
    window.afterExecCommand=true;
    result(function(){
      assert_true(eventFired, eventType + ' event should fire');
      assert_true(execSucceeded, 'document.execCommand() should succeed and return true');
    });
}

    </script>
</head>
<body>

    <p>FAILED (This TC requires JavaScript enabled)</p>
    <iframe id="iframe"></iframe>
    <div id="log"></div>
    <script>
    (function(){
        var t = async_test()
        var dataToPaste='';
        var extPassCond=null;
        var eventTarget='iframe.contentWindow';
        var eventType='copy';
        setup(function(){
                setupTest(t, eventTarget, eventType, dataToPaste, extPassCond );
            });
    })();
    </script>
</body></html>
